Fix broken action menu
